Replace literal octothorpe with HTML entity
OK, sorry to contribute two commits to fix a single character in some documentation. My last solution (#46) worked in the GitHub markdown preview tab, but didn't work when rendering to HTML. I'm not sure why and I'm not going to bother figuring that part out. Using an HTML entity instead of a literal `#` seems like a clearer approach anyways.