Update instructions for using tidal-vis
The old instructions were broken by tidalcycles/Tidal#649, which split up OSC configurations into targets (connection info) and shapes and removed `startMulti`. This change updates the instructions to work with current Tidal.
It also suggests installing tidal-vis from Hackage via cabal rather than building the repo with stack, as the former worked for me where the latter failed. But this might be an issue with stack on my end; let me know if I should revert that.
If this change looks good, I can make a corresponding PR for tidal-vis to update the README (which has the same outdated instructions).